requirements.txt
Keras-Preprocessing==1.1.2
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.4.1
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
spotipy==2.16.0
python-dotenv==0.14.0
pandas==1.1.2
numpy==1.19.2
joblib==0.16.0
fastapi==0.61.1
uvicorn==0.11.8
psycopg2==2.8.6
